move webapp into nightwatch package
